+++ bindings/posix/src/test/scala/org/scalanative/bindings/tests/RegexSpec.scala
@@ -7,6 +7,7 @@ class RegexSpec extends AnyFunSpec {
     it("should match regular expressions") {
       //#usage-example
       import scala.scalanative.unsafe._
+      import scala.scalanative.unsigned._
       import org.scalanative.bindings.posix.regex._
 
       val reg = stackalloc[regex_t]
@@ -15,10 +16,10 @@ class RegexSpec extends AnyFunSpec {
         regcomp(reg, c"Scala \(J\(S\|VM\)\|Native\)", defines.REG_EXTENDED)
       assert(compResult == 0)
 
-      assert(regexec(reg, c"Scala JVM", 0, null, 0) == 0)
-      assert(regexec(reg, c"Scala JS", 0, null, 0) == 0)
-      assert(regexec(reg, c"Scala Native", 0, null, 0) == 0)
-      assert(regexec(reg, c"Scala .NET", 0, null, 0) != 0)
+      assert(regexec(reg, c"Scala JVM", 0.toULong, null, 0) == 0)
+      assert(regexec(reg, c"Scala JS", 0.toULong, null, 0) == 0)
+      assert(regexec(reg, c"Scala Native", 0.toULong, null, 0) == 0)
+      assert(regexec(reg, c"Scala .NET", 0.toULong, null, 0) != 0)
 
       regfree(reg)
       //#usage-example
